Add files via upload
If you would like to see the image displayed in tkinter you will need to add the image you're trying to edit to the root folder of the program. If you just want to have a copy of the gray scale of the image just point it to the location of the image.